Clean up some Mac deprecation warnings
On Macs, apparently, the underlying type of an enum is not int by
default, so fmt complains if you use %d.
The stream stuff I fixed up a bit since that's an ext: function
and we should not be presenting users with internal codes.

+++ src/asttooling/astExpose1.cc
@@ -232,7 +232,7 @@ core::T_sp mostDerivedType(const clang::Type *x) {
   llvm::raw_string_ostream out(buf);
   clang::ASTDumper dumper(out,false);
   dumper.Visit(s);
-  SIMPLE_ERROR(("astExpose.cc>mostDerivedType. Could not cast clang::Type s->getTypeClass()-> %d. typename: %s ") , s->getTypeClass() , out.str() );
+  SIMPLE_ERROR(("astExpose.cc>mostDerivedType. Could not cast clang::Type s->getTypeClass()-> %d. typename: %s ") , (int)(s->getTypeClass()) , out.str() );
 }
